feat: implement support for packages (#56)
Packages are a way to pull in tools from another Gram project. The high level idea is that any project in Gram can enable packages and now other customers will be able to discover these on an upcoming marketplace.

We will leverage packages to build and distribute third party integrations such as HubSpot and Zendesk. Initially, only the Speakeasy organization in Gram will be able to enable packages for projects but support could expand if this idea is successful.