feat: support markdown descriptions for packages (#142)
This PR adds support for setting markdown descriptions on packages. These descriptions are rendered to HTML and sanitized before being stored in the database.

In addition to this change, the PR also adds an /rpc/integrations.get endpoint that will fetch detailed information about an integration which includes the underlying package description.