chore: add http tool override fields to db schema (#169)
* chore: add http tool override fields to db schema

This change adds the following fields to http_tool_definitions table:

- x_gram
- confirm
- confirm_prompt
- original_name
- original_summary
- original_description

We are looking to introduce support for an `x-gram` openapi extension
that allows API owners to customize the name, summary, description and
confirmation prompt for a tool. The additional fields are used to
store the original values for informational purposes.

* update sqlc code

* update goa codegen