chore: ensure otel sdk is always enabled (#177)
This change refactors the opentelemetry sdk to always be enabled so that
we always generate traces and metrics. The main difference with this
change is that when `GRAM_ENABLE_OTEL` is set to 0/false we create
metric and span exporters that discard everything. This is different
than configuring noop metrics/trace providers because those will not
generate any ids.

In short, opentelemetry is always enabled but throws away traces and
metrics by default.