fix: cleanup hard-coded server and dashboard URLs (#260)
* fix: cleanup hard-coded server and dashboard URLs

This change removes the hard-coded server and dashboard URLs from the
codebase that are often mapped by environment and instead introduces
environment variables to configure these as needed. They are
`GRAM_SERVER_URL` for the API server and `GRAM_SITE_URL` for the
dashboard.

Additionally, we now support providing SSL certificates for local
development or if users wish to run the server with SSL enabled. SSL is
required to work with `Secure` cookies during development if the
dashboard is not running on localhost - localhost gets special treatment
by browsers.

The combination of these changes makes it also possible to work with
remote development environments over a virtual network instead of
relying on port forwarding.

* update custom domains middleware