feat: add prompt template rendering endpoint (#279)
* feat: add prompt template rendering endpoint

This change adds a new endpoint to the templates service that allows
clients to render a prompt template with a given set of arguments. It
can be useful to build a preview feature as part of prompt template
building that uses this endpoint to render the prompt template with
test arguments.

* update sdk

* fix sdk clobbering