feat: add metrics around openapi processing (#392)
This change introduces new metrics to monitor OpenAPI processing when
deployments are created. This also includes metrics to detect when Gram
needs to upgrade documents from v3.0 to v3.1. The metrics facility was
refactored to make introducing new metrics easier but to also provide
strongly typed methods instead of giving the rest of the codebase
unrestricted access to metrics.