fix: drop unnecessary unique constraint on http_security table
This commit drops the unnecessary unique constraint on the `key` column
of the `http_security` table. The unique constraint was preventing
deployments with two or more openapi docs that had security schemes with
the same name (key) from being processed successfully. After this
migration there will be multiple rows with the same key but each will
be served by different environment variables as expected.