feat: complete client credentials implementation (#418)
Still doing some testing.

Notable Areas To Review:

- We allow tokenURL to be set either from the openapi config, if
necesarry from the current serverURL, or directly by token URL
environment variable in that heirarchy
- We implement a token cache respecting the expires_in feature of
client_credentials response.
- This cache operates on an exact match of normalizes scopes to ensure
accuracy of the token aligning to the underlying needs of the operation.
(ex: if we have `ledger:read`,`card:read` and we need a token for
`card:read` we will get a new token)
- I think this is an okay solution and won't really cause issues. There
are some other options but they are potentially worse.
- We could implement a cache prefix match that returns multiple options
- This is doable in Redis, but it's by SCAN which is inefficient. There
is also no guarantee other cache type implementation besides redis in
the future support this
- We could keep wonder master cache object that stores many access
tokens for many scopes. This object becomes quite complicated and we can
no longer take advantage of the underlying TTL features of a cache. We
basically have to clean out this object as we go ourselves every time.
It's messy.