chore: rename server go module (#440)
This change change the go module name for the server:

```diff
- module github.com/speakeasy-api/gram
+ module github.com/speakeasy-api/gram/server
```

This is because the gram monorepo will likely host other Go modules such
as the standalone proxy and it will make sense for that the be at
`github.com/ speakeasy-api/gram/proxy` adjacent to the server module.
This will then allow us to potentially import the proxy module in the
server.

Or broadly speaking it will allow multiple Go modules in this monorepo
to import each other using correct module paths.