fix: return a structured method not allowed error (#514)
This change updates the MCP GET endpoint to return a structured json
error representing a method not allowed error instead of a plain text
message. I observed this behavior in a sample repo using Hono to build
an MCP server and it seemed appropriate to be consistent with json-rpc
within this failure mode.


https://github.com/mhart/mcp-hono-stateless/blob/90ea3f09bfb546bad483c5ed0244271f1652df28/src/index.ts#L160