chore: trigger onboarding flow after registration (#526)
This change removes the empty project check that is run on a number of
pages to trigger onboarding. Instead, we simplify this to redirecting to
onboarding after successful registration.

The motivation for this change is to reduce the amount of API calls
being made from the frontend to `/rpc/deployments.latest`.