chore: prevent excess rerendering in tool builder (#529)
The tool builder page was rerendering excessively due to `ChatContext`
wrapping the whole page instead of just the `ChatWindow` component. This
would result in excessive refetching of react query hooks like
`useLatestDeployment` causing a mini-DoS on our servers. Rerendering of
the page was occurring each time a token was rendered - on each call to
`chat.appendMessage`. By moving the `ChatContext` down the tree, we
localize rerendering to the `ChatWindow`.