chore: create environment abstraction access (#540)
This change creates an abstraction layer for accessing environment data
that powers chat and tool execution. The goal in creating the package
`github.com/speakeasy-api/gram/server/internal/gateway` with the
`EnvironmentLoader` interface is to prepare us to extract a standalone
proxy in future work. The proxy will be built with the ability to read
environment variables to power tool calling.