fix: ensure custom HTTP routes are OTel-tagged (#593)
This change ensures that all custom HTTP routes mounted on the server
have an appropriate route tag so they can be tracked with OTel metrics
and traces.

Goa with its OTel plugin automatically does this for service methods. We
need to remember to do it for extra routes we add manually.