feat: generous parsing oauth response (#632)
The official OAuth spec does expect that token responses are exactly
snake case `access_token`. We have a customer that has not implemented
it this way and I expect we may see this again eventually.

I think we can be generous enough to check for camelCase as well here,
but also recognizing we'll need to draw lines somewhere.