fix: ensure all CI jobs run on main branch (#635)
This change updates the pr workflow so that all jobs run on the main
branch, regardless of whether there are changes in the client or server
code. Since branches can be stale on merge, this is necessary to ensure
that the CI checks are always run over the latest changes on main.

Previously certain jobs were being skipped on main.