fix: use more appropriate error status codes than 500 (#655)
This change uses a more appropriate error status code for when a tool
call cannot be made because server url is not set. Previously, we were
sending a 500 status code but that's not appropriate given the server is
explicitly handling the tool call and rejecting it.

The ultimate goal of this change and others that will follow is to
improve our DataDog monitors around http errors. These are currently
very noisy because they are detecting a wide range of 4XX and 5XX
errors. We should trend towards alerting on 5XX errors.