feat: reads for oauth metadata (#650)
Populates data for an `external_oauth_server` or `oauth_proxy_server` in
the toolset details call. This will only be populated for toolset
details **not** list toolsets.

This will allow to show a badge for MCP servers that have either type of
OAuth enabled, alongside some configuration details.

Important to note:

We are not prioritizing doing self service setup of OAuth configurations
right now, so the writes will come from admin scripts. This is for
purely a visualization view.