fix: use ConfigMap to set security headers (#665)
This change moves the security headers configuration from the Ingress
annotations, which had no effect, to a ConfigMap. The ConfigMap is
referenced in the Helm chart to apply the headers correctly.

Additionally, this change also updates the Content-Security-Policy based
on findings from deploying to the dev enviroment. Namely, the pylon chat
widget was being initialized using an inline script tag and we moved it
to a proper script tag to avoid adding 'unsafe-inline' to the CSP
script-src directive.