fix: use codes to represent signin errors (#11)
This change switches the auth redirection to use error codes instead of
error messages. This prevents potential content spoofing attacks because
error messages in the URL represent arbitrary content that an attacker
can control to inject malicious content (like urls) and send to victims.

Now we send back simple error codes and the client side maps the ones it
recognizes to fixed error messages that are shown to the user.