feat: allow referencing secured environments in oauth proxy (#17)
- Allow referencing stored environments for static credentials in oauth
proxy
- Since this is a manual setup. It avoids needing to manually pass any
credentials

There are some downsides here. For example, what if someone deletes an
environment accidentally. Given Oauth proxy is a fairly bespoke setup, I
think we can be ok with this risk for now. Also environments are project
scoped, again I think this is ok. Also adds environments as a dependency
to oauth service, given they are already a dependency to MCP I think
it's ok.