fix: restore openapi download on deployment page (#27)
This change restores the ability to download OpenAPI documents for the
deployment page.

At some point the security for the assets service was tightened to
require the `Gram-Project` header. Since the endpoint
`/rpc/assets.serveOpenAPIv3` is being used as a download link, it is not
passing the project ID. The only thing available was the asset ID in
query params and the gram session cookie which only referenced the
active organization.

To fix this issue, we updated that endpoint to accept an additional,
required `project_id` query parameter and we added a public
`CheckProjectAccess` method to the auth service to verify that a project
id belongs to the active organization. We then amended the security
requirements for the `/rpc/assets.serveOpenAPIv3` endpoint to only
require a valid session or api key and no project slug header.