fix: always add trace.id and span.id attributes
This change updates ContextHandler to always add the `trace.id` and
`span.id` regardless of whether datadog support is enabled or not.

The issue is that datadog deliberately extracts their attributes from
log events such that they are not present in the log record when going
through the log explorer. The UX to surface the hidden trace id is also
very unintuitive. This change gets around that because the vendor
agnostic attributes will be present like any other regular attribute.