chore: reuse CI build of dashboard in docker image (restored)
This commit undoes a recent revert. It was suspected the previous
version of this change was the root cause for a client side outage but
it was not the case.

This change updates our GitHub workflow for PRs and main branch to drop
the docker-build-client job and instead reuse the dashboard build that
happens in the `client-build-lint-test`. The main driver for this change
was that the two builds that happened in CI and Dockerfile had different
outputs for the bundled JS files and therefore different source maps
that could not be correlated by DataDog RUM. This addresses the issue by
copying the CI build into the docker image. It also saves some build
time.