fix: override env vars for vite production builds (#60)
Environment variables in mise.toml are intended for local development.
The CI build of the dashboard was referring to these which would result
in the dashboard not working when deployed. For example, the server url
was set to http://localhost:8080 instead of being
https://{dev,app}.getgram.ai.

This change fixes this issue by create a custom environment for vite
production builds and update the CI pipeline to use it.