fix: do not fail if deployment has no assets (#63)
If all openapi documents are removed in a deployment this would result
in zero tools being created. This currently causes the deployment to be
marked as failed which is incorrect. This change fixes this issue by
allowing deployments with no assets to be marked as completed. If there
are assets in the deployment but no tools were generated, it will still
be marked as failed.