chore: publish docker images for gram functions runtimes
This change introduces the Gram Functions runner project into the monorepo and appropriate CI/CD workflow updates to push runner images to fly.io dev and prod enviroments.

\## What is the runner?
The runner is a fly.io micro-vm (analogous to a docker container) that contains customer-provided JavaScript code and a Go server + JS entrypoint - the Gram Functions runner. The runner's purpose is to receive tool call requests from Gram, authenticate them and spawn a node.js subprocess to execute customer code that returns a result for the tool call, then stream the result back to Gram.

The runner server and entrypoint are built into a docker image and pushed to fly.io (it converts docker images to micro-vm templates).

To build the docker image we are not using the docker CLI. Instead we are using Apko which is a very small tool to build OCI image based on Alpine OS. It takes a list of APK packages (Alpine's equivalent of apt/yum/brew) and generates a super small docker image with an SBOM of everything within it. The Gram Functions runner image with Node.js v22 bundled in comes to ~90MB which allows for very fast deploy and start times in fly.io (<8s to deploy, <300ms to start from stopped state and serve a tool call).

Since Apko requires that everything must be an Alpine APK package, we need a way to package our Go server to be able to include in the image build... That's where melanage comes in. In simplest terms, melange will build a Go package and turn it into Alpine package and then it can be used in Apko builds.

You can inspect the melange build config for Gram Functions runner in functions/melange.config and you can inspect the apko config for the node.js image in functions/images/nodejs22-alpine3.22.yaml.