fix: disable auto-loading env in mise-action (#85)
The recently released mise-action v3 switched automatic loading of env
variables including ones defined in mise.toml. This causes issues with
the dashboard build because now `GRAM_SERVER_URL=http://localhost:8080`
is always set in CI. This is then baked into the vite build of the
dashboard which will fail to communicate with the backend when deploying
to dev/prod environments.