fix: log internal errors in usage service/client (#104)
This change updates the error handling in usage client to return plain
wrapped errors and ensures that consumers of the usage client are
logging these errors appropriately.

Currently, the usage client uses oops.E which results in internal error
messages being masked and only public-facing messages being logged. The
purpose of oops.E is to be used in service layers of the codebase and
not lower levels.