fix: suppress response errors due to request cancelations (#106)
This change updates the logging middleware to coalesce responses to use
non-standard 499 status code if the request context is canceled.

The request context is canceled when the client closes the connection
before the server is finished processing the request. Currently, this
cancelation ends up propagating from a handler as a 500 error that gets
logged and triggers alerts. This isn't desirable because the 500
response will never be sent - the client has already disconnected.

This article does a great job explaining the issue and the reason behind
choosing 499:

https://brandur.org/fragments/testing-request-cancellation