fix: disable skip ci directive in changesets (#196)
The automated version PR that the changesets action creates contains a
`s-kip ci` directive (dash inserted deliberately) which prevents the CI
jobs from running. This in turn means that branch protection rules that
require certain jobs to pass will never be satisfied and the versioning
PR cannot be merged.

This change disables the directive in changesets so that the CI jobs
run.