fix: clear tools from previous deployment attempts
This change adds logic to clear the database of any tools associated
with previous deployment attempts before processing each openapi
document in the deployment.

Before this change, if a deployment successfully processed one document
and failed on another then the whole deployment workflow (Temporal) is
retried. As a result, reprocessing a previously successful document
would result in duplicate tools being created in the database.