fix: revert to zod v3 (#211)
The SDK depends on zod v3 but a recent upgrade to v4 in the dashboard
put the app in a broken state due to dependency confusion. This resulted
in users not being able to get past the login screen.

At this time both the dashboard and SDK need to use zod v3.