chore: call changeset publish to create/push tags (#218)
The changesets action has a `publish` option that points to a script or
command it should use for publishing packages. One requirement that we
missed is that the script must call `changeset publish` at the end. This
would create tags and push them to the git remote.

After this change we are calling `pnpm changeset publish` explicitly
(`pnpm run release` is not needed).