mig: add project_id and openapiv3_document_id to http_security (#226)
This change adds references to projects and deployment openapi documents
to the http_security table. The project_id fkey tends to be useful when
data mining tables with a project in mind. The openapiv3_document_id
fkey is more useful because it helps us identify which document and, in
turn openapi operations, a security scheme applies to.

These new columns must be nullable until we backfill the database.