fix: clear tools from previous deployment attempts (#209)
This change adds logic to clear the database of any http tools and
security associated with previous deployment attempts before processing
each openapi document in the deployment.

Before this change, if a deployment successfully processed one document
and failed on another then the whole deployment workflow (Temporal) is
retried. As a result, reprocessing a previously successful document
would result in duplicate tools being created in the database.