chore: use Gram Bot GitHub App to publish with changesets
This change updates the release workflow to use a GitHub App token when
creating versioning PRs. The reason for this is that CI/CD workflows are
not allowed to run for the default github-actions[bot] represented by
the GITHUB_TOKEN secret. Using a GitHub App installation token solves
this problem.