chore: use Gram Bot GitHub App to publish with changesets (#252)
This change updates the release workflow to use a GitHub App token when
creating versioning PRs. The reason for this is that CI/CD workflows are
not allowed to run for the default github-actions[bot] represented by
the GITHUB_TOKEN secret. Using a GitHub App installation token solves
this problem.


https://docs.github.com/en/actions/how-tos/write-workflows/choose-when-workflows-run/trigger-a-workflow

> When you use the repository's GITHUB_TOKEN to perform tasks, events
triggered by the GITHUB_TOKEN, with the exception of workflow_dispatch
and repository_dispatch, will not create a new workflow run. This
prevents you from accidentally creating recursive workflow runs. For
example, if a workflow run pushes code using the repository's
GITHUB_TOKEN, a new workflow will not run even when the repository
contains a workflow configured to run when push events occur. For more
information, see [Use GITHUB_TOKEN for authentication in
workflows](https://docs.github.com/en/actions/security-guides/automatic-token-authentication).