fix: make sure upgrade CTA shows up (#268)
The reason for this issue was a bit silly. Basically if something isn't
wrapped in a PageSectionCTA, it won't be rendered on the page. For
Billing, we need the upgrade link to be an `a` and not a button, so i
need to change PageSectionCTA to not be a button itself