chore: stop showing summary in UI (#274)
We currently show summary and description for tools in the UI. However,
this is confusing because we don't actually use the summary for
anything. For example, it looks like it should be editable when you go
to edit a tool description, but what actually pops up is _just_ the
description. The summary is entirely uneditable. This PR clarifies
things by simply not showing the summary, which also opens the door to
no longer populating this field