feat: improve MCP install page design and user experience (#295)
# MCP Install Page Improvements

## Summary

Improvements to MCP install page. A lot of this is bringing things
closer to parity with the design. Then also significantly adds 3 more
install targets. Couple more follow ons I'll try to get to today but
this feels like a good shippable increment. Need to finish QAing the
installs to the new targets, but otherwise I think this might be ready
to go.


https://github.com/user-attachments/assets/f2478516-ea7a-453a-b860-4153968b79d0