feat: support uploading Gram Functions as part of deployments (#223)
This change adds support for uploading Gram Functions to load tools
backed by code into Gram. We do not yet have a way to run call these
tools but that is the subject of the next milestone of work on this
project. For now this allows processing functions into tools and
curating them.