fix: openrouter key refresh (#349)
The current openrouter key refresh logic was flawed. Although we set the
credit limit of the key correctly, with the way openrouter works the
usage that is applied to that key never resets.

The best option right now is to create a new fresh key on the monthly
refresh and clean up the old one.

Eventually we've talked about moving this to be centrally metered and
not creating keys per org. We should still eventually move in that
direction, but this is a change that will make our current system behave
correctly.