chore: update tool create/update logic to use urns (#350)
Update toolset creation/updating logic to work off of URNs rather than
raw tool names.

This is a relatively large change, though it is scoped as small as is
technologically possible. This is because we leave around all the old
tool fields used for display in the UI so that only places we write
tools have to change. The changes are:
- listTools now returns httpTools and promptTemplates (conceptually it
should always return all tool types) including URNs
- updateToolset and createToolset have been changed to exclusively
accept a list of tool urns, rather than separate lists of http_tools and
templates
- the prompt_templates relation has been changed to track exclusively
"prompts", e.g. things outright added to the toolset as a prompt (MCP
prompt)
- for backwards compatibility, we still return HOTs in the
prompt_templates field by extracting them from the list of toolURNs
- The custom tools ToolBuilder has been updated to understand/use these
new structures