fix: deflake TestService_UploadFunctions_SupportedContentTypes
This changes fixes a race condition driven by the deduplication logic in
TestService_UploadFunctions_SupportedContentTypes. There were three
parallel subtest that were trying to verify that various zip content
types were allowed in uploadFunction requests. The issue is that they
were all hitting the same service and it was deduplicating the requests
and the serving the upload that was first processed. One of the
assertions in this tests expected the first sub-test to always win
but that was not guaranteed. The fix involves circumventing the
deduplication logic by subtly changing the zip file in each sub-test to
inject a random file. This will result in different SHA256 hashes each
time and each test will result in a full new upload.