feat: connect function tools to toolset concept (#453)
Creates the concept of connection `function_tool_definitions` to
toolsets.
Includes:
- List Tools for Project
- DescribeToolset
- DescribeToolsetEntry
- Gateway model for tools proxy

Adds the FunctionToolDefinition type to the discriminated union "tool"
for Goa design

Generally keeps queries separate between http and function tools to
avoid insanely complicated block queries. We are able to differentiate
which one to use by the tool urn. Discussed this with @chase-crumbaugh

This generally should open up the ability to start integrating function
tools into all the surface areas of the dashboard.
 
_Lots of testing and test writing to be done._