chore: remove authorization from image assets (#493)
I noticed a problem when testing private MCP server install pages.

The problem: Images are currently gated behind session authentication,
but we are using our image storage solution to allow people to publish
logos for their public MCP pages. If a customer shares their install
page, and one of their (unauthenticated) customers goes to look at it,
the image will 401.

The proposal: remove auth from images entirely. It seems to me like it's
reasonable to expect that a gram image is only ever going to be created
for the purpose of branding, which is inherently public content. If we
are to go that route I don't know if we need to make any changes to
product communication. But this seems like a somewhat effortlessly
practical option.

The alternative is implementing a public/private control on images,
which I'm also open to doing but seems like somewhat needless complexity
(though I admit avoids this potential footgun if some strange future
arises in which we are storing sensitive images)

In our current implementation, it should be noted that images are gated
behind authentication but not authorization -- we never check image
ownership. If someone was actually interested in pilfering gram image
assets they would only need to create an account to overcome the
security we have in place right now. If we decide to go the
public/private route, I can double back and add proper authorization for
this case.