chore: release cli when tag already exists (#519)
GoReleaser expects `HEAD` to contain the `@gram/cli` version tag.
However, tags from `changeset` are attached to feature commits, not to
the version bump that’s currently being committed.

This has no effect on external users, who just want to install the
package and start using it. The net effect is that git tags for CLI
versions point at the actual version bump PR, and not to the feature
commit.