feat: mcp install for private servers (#488)
This adds MCP install page for private servers. We accomplish this by
borrowing the cookie from gram application auth. We inherit the
limitation that this cookie will only be available on
https://app.getgram.ai, so custom domain users will not be able to
access their private install page.



couple notes:

* The lookup flow is also somewhat complicated because we don't know
whether auth is required until _after_ the toolset is deemed to have
public MCP or not. For that reason, we do the look up and only check the
session after the toolset is loaded
* The most significant change here is probably moving our cookie Path
from an implicit `/rpc` to an explicit `/`. I cannot imagine this being
a risk in any way, but figured it would be worth calling out. I also
didn't test under which circumstances we reset the cookie, so users who
have a stale cookie path may end up needing to re-login to see their
private servers.
* I added a 307 in the event that a private MCP server is found but
there is no session. This technically leaks the slugs of private
servers. Not sure if we care. We could add a matching 307 if no server
is found at all if we care to
* refactored a fair bit of the logic to make the template rendering
process a bit more linear. I added this SecurityInputs[] concept to make
logic around formatting variables to work as both env vars and http
headers slightly easier to reason about.
* Added some tests around the content negotiation to make sure no
routing breaks. And a test case around not exposing private servers