fix: correctly expire cookies on log out (#549)
Here, we make a couple changes that will improve our cookie behaviors.

First, we were previously passing `MaxAge: 0` to `http.SetCookie` (via
goa). `SetCookie` treats this value as equivalent to unset and so won't
actually expire the cookie. We bypass this by setting MaxAge to -1.

We also remove the `Path` specifier from the DeleteCookie logic. The
idea here is that browers will delete all cookies that match all of the
attributes you specify, but the match need not match _all_ of the
cookies attributes. This lets us ensure that we remove the extraneous
cookies that are stored at `/rpc` without disrupting the system and
achieves the correct behavior with the new cookies.

@ryan-timothy-albert fyi