feat: treat producer keys as consumer superset (#568)
The product documents that we treat producer keys as capable of
performing the read actions of a consumer key. We aren't able to
consistently apply this though. This has caused some user confusion.

This is minorly hacky but helps us reliably implement this intended
behavior. It also keeps us safer should we create more key scopes in the
future.

chosen in favor of this: https://github.com/speakeasy-api/gram/pull/567