chore: implement functions.ToolCaller interface (#570)
This change updates the flyio orchestrator to implement the
`functions.ToolCaller` interface and then updates the gateway proxy to
build requests through that interface. Previously the function tool call
planning coupled directly to flyio which was a temporary measure to get
things working. Now, working through the interface, the gateway has no
knowledge of the underlying orchestrator and allows for swappability and
easier testing.