fix: allow custom install images on custom domains (#578)
In order to achieve this we:

1. Convert custom images to be referenced from the unqualified domain.
`/rpc` paths are not available on custom domains so instead we must use
the gram server URL explicitly.

2. Set `access-control-allow-origin: *` on serveImage requests in order
to allow reading those images from custom domains. There is a world
where we ought to constrain this more tightly, but given our visibility
model on image assets, I think it makes sense. Especially since one need
only set up a free account and add a custom domain in order to
circumvent this protection in a scenario where we tried to allow-origin:
(app.getgram.ai, <some-specific-custom-domain>)