fix: actually ungate auth on images
This realizes the changes attempted in:

https://github.com/speakeasy-api/gram/pull/493

Unfortunately, removing the explicit authorization
in that PR. I'm not thoroughly convinced that this
change is a great idea given that we are now going
to have to opt into authorization for all assets
endpoints. Open to any concept that allows us to
opt serveImage out of auth rather than opting
everything else in.