chore: enable ClickHouse local SSL support (#628)
This PR adds TLS support for the Clickhouse docker container. This
enables Gram communicate with Clickhouse on port 9440 by toggling the
`InsecureSkipVerify` config variable (`false` in prod, `true` in prod),
and avoid messy if blocks when communicating on port 9000 locally and
9440 in prod.

The TLS block is required in to connect to Clickhouse cloud and it's not
supported when connecting on port 9000 (which we used to do locally),
unless we check the port at runtime and choose whether or not to pass
the TLS block to the client.
```go
opts := &clickhouse.Options{
  TLS: &tls.Config{ // can only be included when we're on port 9440
	  InsecureSkipVerify: insecure,
  },
}
```