chore: drop `@gram` scope from package.json names
This change removes the invalid `@gram` scope from the package.json
names. We do not own this scope on npm, so it is misleading to have it.
Though the primary motivation for this PR is that it makes installing
the CLI with aqua and mise slightly simpler.

With aqua this will be reflected as follows:

Before:

```
packages:
  - name: speakeasy-api/gram/cli
    version: "@gram/cli@0.10.0"
```

After:

```
packages:
  - name: speakeasy-api/gram/cli
    version: "cli/0.10.0"
```

Note: we've yet to contribute to the aqua registry because we wanted to
make this naming change first.