feat: add basic toolset UX for resources (#624)
Creates a basic UX for a resources tab within toolsets. 

I considered shoehorning this into the tools view, but given resources
are a separate primitive that we really don't expect to be used very
commonly outside of Apps SDKs examples I think this is better right now.

The UX is more similar to the prompts tab right now. I think eventually
we may want to modify both concepts so they somehow are available in the
centralized tool selector (wouldn't be called that anymore). But I don't
think that's really important right now.