chore: add ts-framework back to workspace (#709)
Motivation here is at least partially to resolve the Vercel build
breakage, which this tragically (and fragilely) fixes by coincidence.
Our dependency `tunnel-rat` depends on react `18.x.x` implicitly and
because this version doesn't otherwise exist in our application, the
rollup build in Vercel chokes on failing to resolve the import. Why
rollup only seeks to traverse this import path in Vercel's build context
is truly beyond me. There's probably an incantation to force tunnel rat
to resolve our version of react, but a more thorough understanding of
what's happening inside Vercel's build process might reveal a better
solution.

Either way, I'm not sure what build breakage motivated this change in
this case, but it was pointed out to me in #645 that we want these
packages to be present in our workspace so we get our friend dependabot
to keep dependencies up to date.